mcp/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
pytest==7.4.2
requests==2.31.0
pyyaml==6.0
httpx==0.24.1